internal/e2e: Move listener tests to featuretests package
Moves the LDS tests from the e2e package to the featuretests package.

Updates #2549
